requirements.txt
openai>=1.54.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=3.0.0
